+++ lib/govuk_publishing_components/presenters/navigation_type.rb
@@ -2,17 +2,13 @@ module GovukPublishingComponents
   module Presenters
     # @private
     class NavigationType
-      GUIDANCE_SCHEMAS =
-        %w{answer contact guide detailed_guide document_collection publication}.freeze
-
       def initialize(content_item)
         @content_item = content_item
       end
 
       def should_present_taxonomy_navigation?
         !content_is_tagged_to_browse_pages? &&
-          content_is_tagged_to_a_live_taxon? &&
-          content_schema_is_guidance?
+          content_is_tagged_to_a_live_taxon?
       end
 
     private